Bugfix powerpad

+++ include/main.h
@@ -45,7 +45,7 @@ uint16_t readZapper();
 // OUTPUT
 void setupBluetooth(void);
 bool connected(void);
-void outputPowerpad(uint8_t powerpadData, uint8_t prevPadData, bool compressed);
+void outputPowerpad(uint16_t powerpadData, uint16_t prevPadData, bool compressed);
 void outputGamepad(uint8_t gamepadData, uint8_t prevPadData);
 void outputZapper(uint8_t zapperData, uint8_t prevPadData);
